+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+### Bug fixes
+
+* [#2169](https://github.com/crypto-org-chain/cronos/pull/2169) fix(cronos): add safe multiply int check for during voucher conversion to evm coins.
+
 *Jul 16, 2026*
 
 ### Chores